# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest API documentation",
        "version": "1"
    },
    "paths": {
        "/api/": {},
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get list of all tasks",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Find by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "deadline",
                                "created_at",
                                "id",
                                "title"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list of all tasks success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskResource"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Create task",
                "operationId": "b263247033817cfd225af5d2a51a93fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "description": "Get task by ID",
                "operationId": "9736e8dac4ca267e196496c4b790c757",
                "parameters": [
                    {
                        "parameter": "task",
                        "name": "task",
                        "in": "path",
                        "description": "Get single taks by ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get task by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TaskIndexRequest": {
                "required": [],
                "properties": {
                    "search": {
                        "type": "nullable|string|max:255",
                        "example": "search by title"
                    },
                    "sort": {
                        "type": "nullable|in:deadline,created_at,id,title",
                        "example": "sorting"
                    },
                    "per_page": {
                        "type": "nullable|integer|min:1|max:100",
                        "example": "paginate"
                    }
                },
                "type": "object"
            },
            "TaskRequest": {
                "required": [],
                "properties": {
                    "title": {
                        "type": "required|string|max:255",
                        "example": "task title"
                    },
                    "description": {
                        "type": "nullable|string",
                        "example": "task description"
                    },
                    "deadline": {
                        "type": "nullable|date",
                        "example": "2025-12-31T23:59:59"
                    },
                    "status": {
                        "type": "nullable|in:pending,completed",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "completed"
                        ]
                    },
                    "priority": {
                        "type": "nullable|in:low,medium,high",
                        "example": "medium",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "category": {
                        "type": "nullable|string|max:100",
                        "example": "task category"
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "properties": {
                    "title": {
                        "type": "required|string|max:255",
                        "example": "task title"
                    },
                    "description": {
                        "type": "nullable|string",
                        "example": "task description"
                    },
                    "deadline": {
                        "type": "nullable|date",
                        "example": "2025-12-31T23:59:59"
                    },
                    "status": {
                        "type": "nullable|in:pending,completed",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "completed"
                        ]
                    },
                    "priority": {
                        "type": "nullable|in:low,medium,high",
                        "example": "medium",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "category": {
                        "type": "nullable|string|max:100",
                        "example": "task category"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "tasks",
            "description": "tasks"
        }
    ]
}